Add files via upload
The main steps followed:
1. Data Analysis -Analyze, summarize and visualize the data and the
interaction between different variables and features.
2. Feature Selection - After the analysis, work on understanding the correlation
between the independent and dependent variables and test with the different model
transformations to figure out which features should be dropped. 
3. Model Selection - Compare the L1 and L2 Norm regularization techniques (Lasso and
Ridge, respectively), different penalty terms, and model transformations to build and
visualize the optimal model. (Use Cross Validation to select the optimal tuning
parameters and compare it with the AIC model selection)
4. Model Evaluation - Divide the data into test and train and evaluate the model
along with fit, residuals, and standard error plots.

Elaboration:-
We started the process with the full model and checked the significance of variables along with the VIF values. The next step was to eliminate variables that were highly correlated such as overall and potential score or were insignificant or had high VIFs. 

We used our intuition to reason that some player statistics (namely, Ball Skills, Physical, Shooting, and Passing) were skewed in ways that favored players who were forwards or midfielders, so we tried different ways to minimize their collective effect. We considered using a Max_Stat variable, which was the maximum of a player’s Ball Skills, Defence, and Goalkeeping scores, in place of the other variables, as well as a Position variable (which took the values Forward, Defender, or Goalkeeper) based on which was their most prominent stat. This would help us see how different positions were favored in the market. We ultimately chose to use Overall Score, as we reasoned that it gave an accurate picture of the player’s ability, and because it was a strong predictor in our early models.

We followed this by transforming the variables based on their relationship with the Market Value. We then had our top three models, which we evaluated using their OOS R^2 and RMSE values. We regularized the best model using Lasso and Ridge regression to reduce the effects of overfitting. Any variables that did not contribute enough to the models’ performance were reduced to zero, and the leftover variables were selected for the final model.

Future Scope
In the future, we may also want to take into account the team a particular player belongs to. This would mean that players of certain teams would receive the correct adjustment for their teams performance, hence overall value. There are also other factors to consider, such as popularity (as football teams stand to make a lot of money from having popular players) as well as recent performance.